优化Windows结构

+++ utils/include/utils/rbtree_base.h
@@ -8,6 +8,8 @@
 #ifndef __RBTREE_H__
 #define __RBTREE_H__
 
+#include <stdint.h>
+
 struct rb_node
 {
     struct rb_node *rb_parent;
@@ -25,7 +27,7 @@ struct rb_root
 
 #define RB_ROOT (struct rb_root){ (struct rb_node *)0, }
 #define rb_entry(ptr, type, member) \
-    ((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+    ((type *)((char *)(ptr)-(uintptr_t)(&((type *)0)->member)))
 
 #ifdef __cplusplus
 extern "C"
